Avoid calling missing `Khan.warnTimeout` function
Summary:
It looks like we got rid of `Khan.warnTimeout` in D24180, back when khan-exercises was still
around, but webapp's KAthJax wasn't updated. (The mobile apps package a separate version of
MathJax and are exhibiting a similar issue -- JS errors due to a missing `Khan.warnTimeout`
function. In the mobile apps, this causes the user to be kicked out of the exercise.)

Test Plan:
In testing these warning functions with both the new and legacy exercise templates,
they don't seem to have any effect (they ultimately trigger a change via a `#warning-bar` element
that doesn't exist). I'm open to removing them altogether, though seems like that should be a
separate change?

Reviewers: emily, james

Reviewed By: emily, james

Differential Revision: https://phabricator.khanacademy.org/D32867